refactor: extractV3Packet()
