+++ marshal.go
@@ -765,7 +765,6 @@ func (x *GoSNMP) unmarshal(packet []byte, response *SnmpPacket) error {
 	if response == nil {
 		return fmt.Errorf("Cannot unmarshal response into nil packet reference")
 	}
-	var OrigMsgFlags = response.MsgFlags
 	var OrigAuthEngineID string
 
 	secParameters, ok := response.SecurityParameters.(*UsmSecurityParameters)
@@ -803,7 +802,13 @@ func (x *GoSNMP) unmarshal(packet []byte, response *SnmpPacket) error {
 			x.Logger.Printf("Parsed version %d", version)
 		}
 	}
-	if response.Version != Version3 {
+
+	if response.Version == Version3 {
+		packet, cursor, err = x.extractV3Packet(packet, cursor, response, OrigAuthEngineID)
+		if err != nil {
+			return err
+		}
+	} else {
 		// Parse community
 		rawCommunity, count, err := x.parseRawField(packet[cursor:], "community")
 		if err != nil {
@@ -816,257 +821,8 @@ func (x *GoSNMP) unmarshal(packet []byte, response *SnmpPacket) error {
 				x.Logger.Printf("Parsed community %s", community)
 			}
 		}
-	} else {
-		if PDUType(packet[cursor]) != Sequence {
-			return fmt.Errorf("Invalid SNMPV3 Header\n")
-		}
-
-		_, cursorTmp := parseLength(packet[cursor:])
-		cursor += cursorTmp
-
-		rawMsgID, count, err := x.parseRawField(packet[cursor:], "msgID")
-		if err != nil {
-			return fmt.Errorf("Error parsing SNMPV3 message ID: %s", err.Error())
-		}
-		cursor += count
-		if MsgID, ok := rawMsgID.(int); ok {
-			response.MsgID = uint32(MsgID)
-			if x.loggingEnabled {
-				x.Logger.Printf("Parsed message ID %d", MsgID)
-
-			}
-		}
-		// discard msg max size
-		_, count, err = x.parseRawField(packet[cursor:], "maxMsgSize")
-		if err != nil {
-			return fmt.Errorf("Error parsing SNMPV3 maxMsgSize: %s", err.Error())
-		}
-		cursor += count
-		// discard msg max size
-
-		rawMsgFlags, count, err := x.parseRawField(packet[cursor:], "msgFlags")
-		if err != nil {
-			return fmt.Errorf("Error parsing SNMPV3 msgFlags: %s", err.Error())
-		}
-		cursor += count
-		if MsgFlags, ok := rawMsgFlags.(string); ok {
-			response.MsgFlags = SnmpV3MsgFlags(MsgFlags[0])
-			if x.loggingEnabled {
-				x.Logger.Printf("parsed msg flags %s", MsgFlags)
-			}
-		}
-
-		rawSecModel, count, err := x.parseRawField(packet[cursor:], "msgSecurityModel")
-		if err != nil {
-			return fmt.Errorf("Error parsing SNMPV3 msgSecModel: %s", err.Error())
-		}
-		cursor += count
-		if SecModel, ok := rawSecModel.(int); ok {
-			response.SecurityModel = SnmpV3SecurityModel(SecModel)
-			if x.loggingEnabled {
-				x.Logger.Printf("Parsed security model %d", SecModel)
-			}
-		}
-
-		if PDUType(packet[cursor]) != OctetString {
-			return fmt.Errorf("Invalid SNMPV3 Security Parameters\n")
-		}
-		_, cursorTmp = parseLength(packet[cursor:])
-		cursor += cursorTmp
-
-		if response.SecurityModel == UserSecurityModel {
-			secParameters, ok := response.SecurityParameters.(*UsmSecurityParameters)
-			if !ok || secParameters == nil {
-				return fmt.Errorf("&GoSNMP.SecurityModel indicates the User Security Model, but &GoSNMP.SecurityParameters is not of type &UsmSecurityParameters")
-			}
-
-			if PDUType(packet[cursor]) != Sequence {
-				return fmt.Errorf("Error parsing SNMPV3 User Security Model parameters\n")
-			}
-			_, cursorTmp = parseLength(packet[cursor:])
-			cursor += cursorTmp
-
-			rawMsgAuthoritativeEngineID, count, err := x.parseRawField(packet[cursor:], "msgAuthoritativeEngineID")
-			if err != nil {
-				return fmt.Errorf("Error parsing SNMPV3 User Security Model msgAuthoritativeEngineID: %s", err.Error())
-			}
-			cursor += count
-			if AuthoritativeEngineID, ok := rawMsgAuthoritativeEngineID.(string); ok {
-				secParameters.AuthoritativeEngineID = AuthoritativeEngineID
-				if x.loggingEnabled {
-					x.Logger.Printf("Parsed authoritativeEngineID %s", AuthoritativeEngineID)
-				}
-			}
-
-			rawMsgAuthoritativeEngineBoots, count, err := x.parseRawField(packet[cursor:], "msgAuthoritativeEngineBoots")
-			if err != nil {
-				return fmt.Errorf("Error parsing SNMPV3 User Security Model msgAuthoritativeEngineBoots: %s", err.Error())
-			}
-			cursor += count
-			if AuthoritativeEngineBoots, ok := rawMsgAuthoritativeEngineBoots.(int); ok {
-				secParameters.AuthoritativeEngineBoots = uint32(AuthoritativeEngineBoots)
-				if x.loggingEnabled {
-					x.Logger.Printf("Parsed authoritativeEngineBoots %d", AuthoritativeEngineBoots)
-				}
-			}
-
-			rawMsgAuthoritativeEngineTime, count, err := x.parseRawField(packet[cursor:], "msgAuthoritativeEngineTime")
-			if err != nil {
-				return fmt.Errorf("Error parsing SNMPV3 User Security Model msgAuthoritativeEngineTime: %s", err.Error())
-			}
-			cursor += count
-			if AuthoritativeEngineTime, ok := rawMsgAuthoritativeEngineTime.(int); ok {
-				secParameters.AuthoritativeEngineTime = uint32(AuthoritativeEngineTime)
-				if x.loggingEnabled {
-					x.Logger.Printf("Parsed authoritativeEngineTime %d", AuthoritativeEngineTime)
-				}
-			}
-
-			rawMsgUserName, count, err := x.parseRawField(packet[cursor:], "msgUserName")
-			if err != nil {
-				return fmt.Errorf("Error parsing SNMPV3 User Security Model msgUserName: %s", err.Error())
-			}
-			cursor += count
-			if msgUserName, ok := rawMsgUserName.(string); ok {
-				secParameters.UserName = msgUserName
-				if x.loggingEnabled {
-					x.Logger.Printf("Parsed userName %s", msgUserName)
-				}
-			}
-
-			rawMsgAuthParameters, count, err := x.parseRawField(packet[cursor:], "msgAuthenticationParameters")
-			if err != nil {
-				return fmt.Errorf("Error parsing SNMPV3 User Security Model msgAuthenticationParameters: %s", err.Error())
-			}
-
-			if msgAuthenticationParameters, ok := rawMsgAuthParameters.(string); ok {
-				secParameters.AuthenticationParameters = msgAuthenticationParameters
-				if x.loggingEnabled {
-					x.Logger.Printf("Parsed authenticationParameters %s", msgAuthenticationParameters)
-				}
-			}
-			// use the authoritative copy of MsgFlags to determine whether this message should be authenticated
-			if OrigMsgFlags&AuthNoPriv > 0 {
-				if count != 14 {
-					return fmt.Errorf("Error authenticating incoming packet: msgAuthenticationParameters is not the correct size")
-				}
-				blank := make([]byte, 12)
-				copy(packet[cursor+2:cursor+14], blank)
-				// secParameters.AuthenticationProtocol and secParameters.AuthenticationPassphrase are not written to in this function,
-				// so no need to save an 'original' to authenticate against
-				if !isAuthentic(packet, secParameters.AuthenticationParameters,
-					secParameters.AuthenticationProtocol,
-					secParameters.AuthenticationPassphrase,
-					OrigAuthEngineID) {
-					return fmt.Errorf("Incoming packet is not authentic, discarding")
-				}
-			}
-			cursor += count
-
-			rawMsgPrivacyParameters, count, err := x.parseRawField(packet[cursor:], "msgPrivacyParameters")
-			if err != nil {
-				return fmt.Errorf("Error parsing SNMPV3 User Security Model msgPrivacyParameters: %s", err.Error())
-			}
-			cursor += count
-			if msgPrivacyParameters, ok := rawMsgPrivacyParameters.(string); ok {
-				secParameters.PrivacyParameters = []byte(msgPrivacyParameters)
-				if x.loggingEnabled {
-					x.Logger.Printf("Parsed privacyParameters %s", msgPrivacyParameters)
-				}
-			}
-
-			//response.SecurityParameters = &secParameters
-		}
-		switch PDUType(packet[cursor]) {
-		case OctetString:
-			// pdu is encrypted
-			_, cursorTmp := parseLength(packet[cursor:])
-			cursorTmp += cursor
-
-			if response.SecurityModel == UserSecurityModel {
-				var secParams *UsmSecurityParameters
-				secParams, ok := response.SecurityParameters.(*UsmSecurityParameters)
-				if !ok || secParams == nil {
-					return fmt.Errorf("response.SecurityModel indicates the User Security Model, but response.SecurityParameters is not of type &UsmSecurityParameters")
-				}
-				var privkey = genlocalkey(secParams.AuthenticationProtocol,
-					secParams.PrivacyPassphrase,
-					secParams.AuthoritativeEngineID)
-				switch secParams.PrivacyProtocol {
-				case AES:
-					var iv [16]byte
-					binary.BigEndian.PutUint32(iv[:], secParams.AuthoritativeEngineBoots)
-					binary.BigEndian.PutUint32(iv[4:], secParams.AuthoritativeEngineTime)
-					copy(iv[8:], secParams.PrivacyParameters)
-
-					block, err := aes.NewCipher(privkey[:16])
-					if err != nil {
-						return err
-					}
-					stream := cipher.NewCFBDecrypter(block, iv[:])
-					plaintext := make([]byte, len(packet[cursorTmp:]))
-					stream.XORKeyStream(plaintext, packet[cursorTmp:])
-					copy(packet[cursor:], plaintext)
-					packet = packet[:cursor+len(plaintext)]
-				default:
-					if len(packet[cursorTmp:])%des.BlockSize != 0 {
-						return fmt.Errorf("Error decrypting ScopedPDU: not multiple of des block size.")
-					}
-					preiv := privkey[8:]
-					var iv [8]byte
-					for i := 0; i < len(iv); i++ {
-						iv[i] = preiv[i] ^ secParams.PrivacyParameters[i]
-					}
-					block, err := des.NewCipher(privkey[:8])
-					if err != nil {
-						return err
-					}
-					mode := cipher.NewCBCDecrypter(block, iv[:])
-
-					plaintext := make([]byte, len(packet[cursorTmp:]))
-					mode.CryptBlocks(plaintext, packet[cursorTmp:])
-					copy(packet[cursor:], plaintext)
-					// truncate packet to remove extra space caused by the
-					// octetstring/length header that was just replaced
-					packet = packet[:cursor+len(plaintext)]
-				}
-
-			}
-			fallthrough
-		case Sequence:
-			// pdu is plaintext
-			tlength, cursorTmp := parseLength(packet[cursor:])
-			// truncate padding that may have been included with
-			// the encrypted PDU
-			packet = packet[:cursor+tlength]
-			cursor += cursorTmp
-			rawContextEngineID, count, err := x.parseRawField(packet[cursor:], "contextEngineID")
-			if err != nil {
-				return fmt.Errorf("Error parsing SNMPV3 contextEngineID: %s", err.Error())
-			}
-			cursor += count
-			if contextEngineID, ok := rawContextEngineID.(string); ok {
-				response.ContextEngineID = contextEngineID
-				if x.loggingEnabled {
-					x.Logger.Printf("Parsed contextEngineID %s", contextEngineID)
-				}
-			}
-			rawContextName, count, err := x.parseRawField(packet[cursor:], "contextName")
-			if err != nil {
-				return fmt.Errorf("Error parsing SNMPV3 contextName: %s", err.Error())
-			}
-			cursor += count
-			if contextName, ok := rawContextName.(string); ok {
-				response.ContextName = contextName
-				if x.loggingEnabled {
-					x.Logger.Printf("Parsed contextName %s", contextName)
-				}
-			}
-
-		default:
-			return fmt.Errorf("Error parsing SNMPV3 scoped PDU\n")
-		}
 	}
+
 	// Parse SNMP packet type
 	requestType := PDUType(packet[cursor])
 	switch requestType {
